[#58] remove events from nuxt plugins
