+++ nuxt.config.ts
@@ -43,7 +43,6 @@ export default defineNuxtConfig({
   css: ["~/assets/index.css"],
   plugins: [
     {src: '~/plugins/api.client.ts'},
-    {src: '~/plugins/events.client.ts'},
     {src: '~/plugins/vendors.client.ts'},
   ],
   modules: [
