+++ src/shared/lib/use-events/use-events.ts
@@ -1,10 +1,14 @@
 import { storeToRefs } from "pinia";
 import type { Ref } from "vue";
-import type { ServerEvent, NormalizedEvent } from '../../types';
+import type { RayContentLock } from "~/src/entities/ray/types";
+import type { ServerEvent, NormalizedEvent, EventId } from '../../types';
+import { useApiTransport } from "../use-api-transport";
 import { normalizeUnknownEvent } from "./normalize-unknown-event";
 import { useEventsPlugin, type TUseEventsPluginData } from "./use-events-plugin";
 import type { TCachedEventsEmptyMap, TEventsGroup } from "~/stores/cached-ids";
 import { useCachedIdsStore } from "~/stores/cached-ids";
+import { useLockedIdsStore } from "~/stores/locked-ids";
+
 
 type TUseEvents = () => {
   normalizeUnknownEvent: (event: ServerEvent<unknown>) => NormalizedEvent<unknown>
@@ -13,15 +17,34 @@ type TUseEvents = () => {
     idsByType: Ref<TCachedEventsEmptyMap>;
     stopUpdatesByType: (type: TEventsGroup) => void
     runUpdatesByType: (type: TEventsGroup) => void
+  },
+  lockedIds: {
+    items: Ref<EventId[]>
+    add: (id: EventId) => void
+    remove: (id: EventId) => void
+  },
+  rayExecution: {
+    continue: (contentName: RayContentLock["name"]) => void
+    stop: (contentName: RayContentLock["name"]) => void
   }
 }
 
 export const useEvents: TUseEvents = () => {
   const cachedIdsStore = useCachedIdsStore();
+  const lockedIdsStore = useLockedIdsStore();
+
+  const {
+    rayContinueExecution,
+    rayStopExecution,
+  } = useApiTransport();
 
   const {
     cachedIds,
   } = storeToRefs(cachedIdsStore)
+
+  const {
+    lockedIds,
+  } = storeToRefs(lockedIdsStore)
 
   return {
     normalizeUnknownEvent,
@@ -31,5 +54,14 @@ export const useEvents: TUseEvents = () => {
       stopUpdatesByType: cachedIdsStore.setByType,
       runUpdatesByType: cachedIdsStore.removeByType,
     },
+    lockedIds: {
+      items: lockedIds as unknown as Ref<EventId[]>,
+      add: lockedIdsStore.add,
+      remove: lockedIdsStore.remove,
+    },
+    rayExecution: {
+      continue: rayContinueExecution,
+      stop: rayStopExecution,
+    }
   }
 }
